+++ spec/simplecov/tdd_spec.rb
@@ -1,17 +1,18 @@
 RSpec.describe SimpleCov::Tdd do
 
   let(:formatter) { SimpleCov::Formatter::Tdd.new }
-  let(:matched_90_filename) { Dir.pwd + "/app/models/matched_90.rb" }
+  let(:matched_90_filename) { "app/models/matched_90.rb" }
+  let(:matched_90_filepath) { Dir.pwd + "/" + matched_90_filename }
   let(:matched_90_spec) { "spec/models/matched_90_spec.rb" }
   let(:matched_file_90_coverage) do
     instance_double(
       SimpleCov::SourceFile,
-      filename: matched_90_filename,
+      filename: matched_90_filepath,
       covered_percent: 90.0,
       lines_of_code: 167,
       missed_lines: [
-        instance_double(SimpleCov::SourceFile::Line, line_number: 5),
-        instance_double(SimpleCov::SourceFile::Line, line_number: 25)
+        instance_double(SimpleCov::SourceFile::Line, line_number: 5, src: "    obj.is_a?(SomeClass)"),
+        instance_double(SimpleCov::SourceFile::Line, line_number: 25, src: "    SomeClass.explode!")
       ]
     )
   end
@@ -27,10 +28,22 @@ RSpec.describe SimpleCov::Tdd do
     )
   end
   let(:unmatched_file) { instance_double(SimpleCov::SourceFile, filename: "spec/models/narp_spec.rb") }
+  let(:matched_file_with_intersecting_filename) { "app/models/anotherspec_with_spec.rb" }
+  let(:matched_file_with_intersecting_filepath) { Dir.pwd + "/" + matched_file_with_intersecting_filename }
+  let(:matched_file_with_intersecting_spec) { "spec/models/anotherspec_with_spec_spec.rb" }
+  let(:matched_file_with_intersecting_name_source_file) do
+    instance_double(
+      SimpleCov::SourceFile,
+      filename: matched_file_with_intersecting_filepath,
+      covered_percent: 100.0,
+      lines_of_code: 53,
+      missed_lines: []
+    )
+  end
   let(:simple_cov_result_stub) do
     instance_double(
       SimpleCov::Result,
-      files: [matched_file_90_coverage, unmatched_file, matched_file_100_coverage]
+      files: [matched_file_90_coverage, unmatched_file, matched_file_100_coverage, matched_file_with_intersecting_name_source_file]
     )
   end
 
@@ -53,9 +66,10 @@ RSpec.describe SimpleCov::Tdd do
 
         it "raises an Exception" do
           stub_const("ARGV", ["misc", "nope", "2", "spec/unfound_spec.rb"])
+          SimpleCov::Formatter::Tdd.debug_mode = true
 
           expect {
-            formatter.format(simple_cov_result_stub, debug: true)
+            formatter.format(simple_cov_result_stub)
           }.to raise_error(StandardError)
         end
       end
@@ -63,7 +77,7 @@ RSpec.describe SimpleCov::Tdd do
 
     context "with a matched filename" do
 
-      context "and without full 100% coverage" do
+      context "without full 100% coverage" do
         it "displays the coverage overview" do
           stub_const("ARGV", ["misc", "nope", "2", matched_90_spec])
 
@@ -88,9 +102,23 @@ RSpec.describe SimpleCov::Tdd do
             formatter.format(simple_cov_result_stub)
           }.to output(/#{line_number_array}/).to_stdout
         end
+
+        context "and output_style is :verbose" do
+
+          it "displays the line numbers along with matching code" do
+            stub_const("ARGV", ["misc", "nope", "2", matched_90_spec])
+            SimpleCov::Formatter::Tdd.output_style = :verbose
+
+            expect {
+              formatter.format(simple_cov_result_stub)
+            }.to output(
+              /line \| source code\n-------------------\n#{matched_file_90_coverage.missed_lines.first.line_number} => #{Regexp.quote(matched_file_90_coverage.missed_lines.first.src.strip)}\n#{matched_file_90_coverage.missed_lines[1].line_number} => #{Regexp.quote(matched_file_90_coverage.missed_lines[1].src.strip)}\n/
+            ).to_stdout
+          end
+        end
       end
 
-      context "and with full 100% coverage" do
+      context "with full 100% coverage" do
         it "displays the coverage overview" do
           stub_const("ARGV", ["misc", "nope", "2", matched_100_spec])
 
@@ -109,6 +137,14 @@ RSpec.describe SimpleCov::Tdd do
       end
 
       context "with different command line argument orders" do
+        it "displays the matched filename when first argument" do
+          stub_const("ARGV", [matched_90_spec, "misc", "nope", "2"])
+
+          expect {
+            formatter.format(simple_cov_result_stub)
+          }.to output(/#{matched_90_filename}/).to_stdout
+        end
+
         it "displays the matched filename when middle argument" do
           stub_const("ARGV", ["misc", matched_90_spec, "nope", "2"])
 
@@ -125,11 +161,17 @@ RSpec.describe SimpleCov::Tdd do
           }.to output(/#{matched_90_filename}/).to_stdout
         end
       end
+
+      context "when filename contains the phrase 'spec'" do
+
+        it "displays the matched filename" do
+          stub_const("ARGV", [matched_file_with_intersecting_spec, "misc", "nope", "2"])
+
+          expect {
+            formatter.format(simple_cov_result_stub)
+          }.to output(/#{matched_file_with_intersecting_filename}/).to_stdout
+        end
+      end
     end
   end
-
-  # @todo green coverage
-  # @todo brown coverage
-  # @todo red coverage
-  # @todo tty color coverage
 end